pyproject.toml
[project]
name = "blockvault"
version = "0.1.0"
description = "BlockVault Phase 1 - Flask backend with MetaMask authentication via signature verification"
authors = [ { name = "ScriptXDemon" } ]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "Flask>=3.0.0,<4.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "ipfshttpclient>=0.8.0a2,<1.0.0",
  "Flask-Cors>=4.0.0,<5.0.0",
  "cryptography>=42.0.0,<43.0.0",
  "web3>=6.20.0,<7.0.0",
]

[project.optional-dependencies]
dev = ["pytest >=8.0.0,<9.0.0", "black >=24.0.0,<25.0.0", "flake8 >=7.0.0,<8.0.0", "mypy >=1.11.0,<2.0.0", "types-PyJWT", "types-requests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["blockvault"]

[tool.black]
line-length = 100

[tool.flake8]
max-line-length = 100
exclude = [".venv", "build", "dist", "__pycache__"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false
